build: move pure function out to ease testing
