+++ src/index.tsx
@@ -5,4 +5,4 @@ export { UserInterface } from "./ui";
 export type { Spline } from "@/toolboxes/spline";
 export type { BoundingBox } from "@/toolboxes/boundingBox";
 export type { BrushStroke } from "@/toolboxes/paintbrush";
-export { evaluateBezier } from "@/toolboxes/spline/Canvas";
+export { evaluateBezier } from "./evaluateBezier";
